Moved the run scenario component to what-if module
