+++ ui/src/app/app.module.ts
@@ -24,7 +24,7 @@ import { GoogleChartsModule } from 'angular-google-charts';
 import { WhatIfComponent } from './modules/what-if/components/whatif/what-if.component';
 import { MigrateComponent } from './migrate/migrate.component';
 import { FormsModule } from '@angular/forms';
-import { RunScenarioComponent } from './run-scenario/run-scenario.component';
+import { RunScenarioComponent } from './modules/what-if/components/run-scenario/run-scenario.component';
 import { WhatIfModule } from './modules/what-if/what-if.module';
 
 @NgModule({
